auth_service/requirements.txt
grpcio
grpcio-tools
pymongo
passlib[bcrypt]
PyJWT
python-dotenv
